Transaction edit data
