+++ FinancialAssistantApp/app/src/main/java/com/coderaptor/financial/assistant/app/MainActivity.kt
@@ -3,12 +3,8 @@ package com.coderaptor.financial.assistant.app
 import android.annotation.SuppressLint
 import android.content.Intent
 import android.os.Bundle
-import android.text.Editable
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
-import com.afollestad.materialdialogs.MaterialDialog
-import com.afollestad.materialdialogs.customview.customView
-import com.afollestad.materialdialogs.customview.getCustomView
 import com.afollestad.recyclical.datasource.DataSource
 import com.afollestad.recyclical.datasource.dataSourceOf
 import com.afollestad.recyclical.setup
@@ -26,10 +22,10 @@ import com.coderaptor.financial.assistant.app.features.sms.askPermission
 import com.coderaptor.financial.assistant.app.features.sms.getSmsMessages
 import com.coderaptor.financial.assistant.app.gui.DreamActivity
 import com.coderaptor.financial.assistant.app.gui.RepeatActivity
-import com.coderaptor.financial.assistant.app.util.*
+import com.coderaptor.financial.assistant.app.gui.openOnceTransactionDialog
+import com.coderaptor.financial.assistant.app.util.SharedPreference
+import com.coderaptor.financial.assistant.app.util.formatDate
 import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.dialog_add_income.*
-import kotlinx.android.synthetic.main.dialog_add_income.view.*
 import java.util.*
 
 
@@ -59,47 +55,7 @@ class MainActivity : AppCompatActivity(){
         val dataSource: DataSource<Any> = dataSourceOf(getOneWeekData(dbHandler))
 
         addNewButton.setOnClickListener {
-            MaterialDialog(this).show {
-                setTheme(R.style.AppTheme)
-                title(R.string.onceNewtransaction)
-                customView(R.layout.dialog_add_income, scrollable = true)
-
-                val datefield = getCustomView().dateField
-                datefield.isClickable = true
-                datefield.text = Editable.Factory.getInstance().newEditable(java.util.Calendar.getInstance().formatDate())
-                datefield.setOnClickListener {
-                    openCalendar(it.dateField)
-                }
-
-                positiveButton(R.string.save) {
-                    val result = fieldsEmpty(amountField.text)
-
-                    if (result){
-
-                        var amount: Int = amountField.text.toString().toInt()
-                        if (kiadas.isChecked) amount = amountField.text.toString().toInt() * -1
-                        val date = dateField.text.toString()
-                        val category: String = categoryField.selectedItem.toString()
-                        val comment = descriptField.text.toString()
-                        var transaction = Transaction(amount, date, category)
-                        if (comment.isNotEmpty()) {
-                            transaction = Transaction(amount, date, comment, category)
-                        }
-
-                        dbHandler.insert(transaction)
-                        dataSource.add(transaction)
-                        if (dbHandler.getCurrentLimit() < 0) {
-                            toast("Napi limit összeg meghaladva")
-                        }
-
-                        toast("sikeres hozzáadás")
-                        }
-                        else{
-                            toast("Hiányzó adat!")
-                        }
-                    }
-                negativeButton(R.string.cancel)
-            }
+            openOnceTransactionDialog(dataSource, dbHandler)
         }
 
         repeatButton.setOnClickListener {
@@ -138,7 +94,6 @@ class MainActivity : AppCompatActivity(){
                 text(R.string.delete)
                 color(R.color.delete)
                 callback { index, item ->
-                    //toast("delete $index: ${item}")
                     if (item is Transaction) {
                         dbHandler.deleteByPosition(item.id, DatabaseHandler.TABLE_NAME_TRANSACTION)
                     } else if (item is Receipt) {
@@ -153,9 +108,11 @@ class MainActivity : AppCompatActivity(){
                 text(R.string.edit)
                 color(R.color.edit)
                 callback { index, item ->
-                    //toast("edit $index: ${item}")
                     if (item is Transaction) {
-                        //edit layout
+                        if (!item.hasFrequency()) {
+                            Log.i("dialog", "hass: true")
+                            openOnceTransactionDialog(dataSource, dbHandler, item)
+                        }
                     } else if (item is Receipt) {
                         val intent = Intent(this@MainActivity, ReceiptActivity::class.java)
                         intent.putExtra("id", item.baseID)
